Implemented the OpenEndedQuiz , remains to fix some minor bugs and fix the intruction and parsing of the response

+++ app/src/main/java/com/example/chatbot/main/data/database_questions/entity/Question.kt
@@ -6,7 +6,7 @@ import com.example.chatbot.main.data.database_messages.model.QuizMetadata
 
 @Entity(tableName = "question_table")
 class Question(@PrimaryKey(autoGenerate = false)
-               val  uid:Int = -1,
+               val  uid:Long = -1,
                val questionUid:Int  = -1,
                val questionContent:String= "",
                val correctAnswer: String = "",
